Move functions to slice folders.
